Refactor apidb tests to be in separate files for better readability.
