+++ src/Makefile.am
@@ -12,10 +12,17 @@ ___test_test_core_LDADD=
 
 if ENABLE_APIDB
 lib_LTLIBRARIES+=libcgimap_apidb.la
-check_PROGRAMS+=../test/test_apidb_backend
+check_PROGRAMS+=\
+	../test/test_apidb_backend_nodes \
+	../test/test_apidb_backend_oauth \
+	../test/test_apidb_backend_historic \
+	../test/test_apidb_backend_changesets
 ___openstreetmap_cgimap_LDADD+=libcgimap_apidb.la
 ___test_test_core_LDADD+=libcgimap_apidb.la
-___test_test_apidb_backend_LDADD=libcgimap_core.la libcgimap_apidb.la
+___test_test_apidb_backend_nodes_LDADD=libcgimap_core.la libcgimap_apidb.la
+___test_test_apidb_backend_oauth_LDADD=libcgimap_core.la libcgimap_apidb.la
+___test_test_apidb_backend_historic_LDADD=libcgimap_core.la libcgimap_apidb.la
+___test_test_apidb_backend_changesets_LDADD=libcgimap_core.la libcgimap_apidb.la
 endif
 
 if ENABLE_PGSNAPSHOT
@@ -23,7 +30,10 @@ lib_LTLIBRARIES+=libcgimap_pgsnapshot.la
 ___openstreetmap_cgimap_LDADD+=libcgimap_pgsnapshot.la
 ___test_test_core_LDADD+=libcgimap_pgsnapshot.la
 if ENABLE_APIDB
-___test_test_apidb_backend_LDADD+=libcgimap_pgsnapshot.la
+___test_test_apidb_backend_nodes_LDADD+=libcgimap_pgsnapshot.la
+___test_test_apidb_backend_oauth_LDADD+=libcgimap_pgsnapshot.la
+___test_test_apidb_backend_historic_LDADD+=libcgimap_pgsnapshot.la
+___test_test_apidb_backend_changesets_LDADD+=libcgimap_pgsnapshot.la
 endif
 endif
 
@@ -31,7 +41,10 @@ lib_LTLIBRARIES+=libcgimap_staticxml.la
 ___openstreetmap_cgimap_LDADD+=libcgimap_staticxml.la
 ___test_test_core_LDADD+=libcgimap_staticxml.la
 if ENABLE_APIDB
-___test_test_apidb_backend_LDADD+=libcgimap_staticxml.la
+___test_test_apidb_backend_nodes_LDADD+=libcgimap_staticxml.la
+___test_test_apidb_backend_oauth_LDADD+=libcgimap_staticxml.la
+___test_test_apidb_backend_historic_LDADD+=libcgimap_staticxml.la
+___test_test_apidb_backend_changesets_LDADD+=libcgimap_staticxml.la
 endif
 
 ___openstreetmap_cgimap_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
@@ -43,7 +56,10 @@ ___test_test_http_LDADD=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FIL
 ___test_test_parse_time_LDADD=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_DATE_TIME_LIB@ @BOOST_SYSTEM_LIB@ @LIBPQXX_LIBS@
 
 if ENABLE_APIDB
-___test_test_apidb_backend_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @BOOST_LOCALE_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
+___test_test_apidb_backend_nodes_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @BOOST_LOCALE_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
+___test_test_apidb_backend_oauth_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @BOOST_LOCALE_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
+___test_test_apidb_backend_historic_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @BOOST_LOCALE_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
+___test_test_apidb_backend_changesets_LDADD+=libcgimap_core.la @BOOST_PROGRAM_OPTIONS_LIB@ @BOOST_FILESYSTEM_LIB@ @BOOST_SYSTEM_LIB@ @BOOST_LOCALE_LIB@ @LIBPQXX_LIBS@ @CRYPTOPP_LIBS@
 endif
 
 ################################################################################
@@ -67,8 +83,23 @@ ___test_test_parse_time_SOURCES=\
 	../test/test_parse_time.cpp
 
 if ENABLE_APIDB
-___test_test_apidb_backend_SOURCES=\
-	../test/test_apidb_backend.cpp \
+___test_test_apidb_backend_nodes_SOURCES=\
+	../test/test_apidb_backend_nodes.cpp \
+	../test/test_formatter.cpp \
+	../test/test_database.cpp \
+	../test/test_request.cpp
+___test_test_apidb_backend_oauth_SOURCES=\
+	../test/test_apidb_backend_oauth.cpp \
+	../test/test_formatter.cpp \
+	../test/test_database.cpp \
+	../test/test_request.cpp
+___test_test_apidb_backend_historic_SOURCES=\
+	../test/test_apidb_backend_historic.cpp \
+	../test/test_formatter.cpp \
+	../test/test_database.cpp \
+	../test/test_request.cpp
+___test_test_apidb_backend_changesets_SOURCES=\
+	../test/test_apidb_backend_changesets.cpp \
 	../test/test_formatter.cpp \
 	../test/test_database.cpp \
 	../test/test_request.cpp
